:sparkles: Slider manager
