+++ mymocore/Core/resources/lang/en/app.php
@@ -164,8 +164,6 @@ return [
   'choose_menu' => 'Choose menu',
   'choose_page' => 'Choose page',
   'open_new_tab' => 'Open new tab',
-  'are_you_sure_you_want_to_delete_this_banner' => 'Are you sure you want to delete this banner?',
-  'sliders' => 'Sliders',
   'link' => 'Link',
   'add_new_banner' => 'Add new banner',
   'change_photo' => 'Change photo',
