+++ plugins/Movie/resources/lang/en/app.php
@@ -138,4 +138,6 @@ return [
     'url' => 'URL',
     'label' => 'Label',
     'action' => 'Action',
+    'are_you_sure_you_want_to_delete_this_banner' => 'Are you sure you want to delete this banner?',
+    'sliders' => 'Sliders',
 ];
