+++ plugins/Movie/routes/backend/backend.route.php
@@ -4,48 +4,6 @@ require_once __DIR__ . '/components/tmdb.route.php';
 require_once __DIR__ . '/components/live-tv.route.php';
 //require_once __DIR__ . '/components/server-stream.route.php';
 
-Route::group(['prefix' => 'genres'], function () {
-    Route::get('/', 'Backend\GenresController@index')->name('admin.genres');
-    
-    Route::get('/getdata', 'Backend\GenresController@getData')->name('admin.genres.getdata');
-    
-    Route::get('/create', 'Backend\GenresController@form')->name('admin.genres.create');
-
-    Route::get('/edit/{id}', 'Backend\GenresController@form')->name('admin.genres.edit')->where('id', '[0-9]+');
-    
-    Route::post('/save', 'Backend\GenresController@save')->name('admin.genres.save');
-    
-    Route::post('/remove', 'Backend\GenresController@remove')->name('admin.genres.remove');
-});
-
-Route::group(['prefix' => 'types'], function () {
-    Route::get('/', 'Backend\TypesController@index')->name('admin.types');
-    
-    Route::get('/getdata', 'Backend\TypesController@getData')->name('admin.types.getdata');
-    
-    Route::get('/create', 'Backend\TypesController@form')->name('admin.types.create');
-    
-    Route::get('/edit/{id}', 'Backend\TypesController@form')->name('admin.types.edit')->where('id', '[0-9]+');
-    
-    Route::post('/save', 'Backend\TypesController@save')->name('admin.types.save');
-    
-    Route::post('/remove', 'Backend\TypesController@remove')->name('admin.types.remove');
-});
-
-Route::group(['prefix' => 'countries'], function () {
-    Route::get('/', 'Backend\CountriesController@index')->name('admin.countries');
-    
-    Route::get('/getdata', 'Backend\CountriesController@getData')->name('admin.countries.getdata');
-    
-    Route::get('/create', 'Backend\CountriesController@form')->name('admin.countries.create');
-
-    Route::get('/edit/{id}', 'Backend\CountriesController@form')->name('admin.countries.edit')->where('id', '[0-9]+');
-    
-    Route::post('/save', 'Backend\CountriesController@save')->name('admin.countries.save');
-    
-    Route::post('/remove', 'Backend\CountriesController@remove')->name('admin.countries.remove');
-});
-
 Route::group(['prefix' => 'stars'], function () {
     Route::get('/', 'Backend\StarsController@index')->name('admin.stars');
     
@@ -60,17 +18,4 @@ Route::group(['prefix' => 'stars'], function () {
     Route::post('/remove', 'Backend\StarsController@remove')->name('admin.stars.remove');
 });
 
-Route::group(['prefix' => 'sliders'], function () {
-    Route::get('/', 'Backend\Design\SlidersController@index')->name('admin.design.sliders');
-
-    Route::get('/getdata', 'Backend\Design\SlidersController@getData')->name('admin.design.sliders.getdata');
-
-    Route::get('/create', 'Backend\Design\SlidersController@form')->name('admin.design.sliders.create');
-
-    Route::get('/edit/{id}', 'Backend\Design\SlidersController@form')->name('admin.design.sliders.edit')->where('id', '[0-9]+');
-
-    Route::post('/save', 'Backend\Design\SlidersController@save')->name('admin.design.sliders.save');
-
-    Route::post('/remove', 'Backend\Design\SlidersController@remove')->name('admin.design.sliders.remove');
-});
