+++ plugins/Movie/routes/web.php
@@ -107,6 +107,8 @@ Route::group(['prefix' => config('mymo_core.admin_prefix'), 'middleware' => ['we
 
     Route::postTypeResource('tv-series', 'Backend\TVSerieController');
 
+    Route::mymoResource('sliders', 'Backend\SliderController');
+
     require_once __DIR__ . '/backend/components/tmdb.route.php';
 });
